pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pairwise-llm-eval"
version = "0.1.0"
description = "Blinded pairwise LLM evaluation with position bias controls and statistical rigor"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "Teal Maker Consulting"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["numpy>=1.24.0", "scipy>=1.10.0", "requests>=2.31.0"]

[project.optional-dependencies]
database = ["psycopg2-binary>=2.9.0"]
dev = ["pytest>=7.0", "responses>=0.23.0"]

[project.scripts]
pairwise-llm-eval = "pairwise_llm_eval.cli:main"

[project.urls]
Homepage = "https://github.com/Teal-Maker/pairwise-llm-eval"

[tool.hatch.build.targets.wheel]
packages = ["src/pairwise_llm_eval"]
artifacts = ["src/pairwise_llm_eval/general_questions.json"]

[tool.hatch.build.targets.sdist]
exclude = [".venv/", "dist/", "*.pyc", "uv.lock"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
